removed more and more from init.lua into filetypes that made more sense
