+++ init.lua
@@ -17,6 +17,10 @@ if not vim.loop.fs_stat(lazypath) then
 end
 vim.opt.rtp:prepend(lazypath)
 
+require('config.options')
+require('config.keymaps')
+require('config.autocmds')
+
 --  You can also configure plugins after the setup call,
 --  as they will be available in your neovim runtime.
 require('lazy').setup({
@@ -165,77 +169,6 @@ require('lazy').setup({
   -- Load plugins defined in custom.plugins
   { import = 'custom.plugins' },
 }, {})
-
--- [[ Setting options ]]
--- See `:help vim.o`
--- NOTE: You can change these options as you wish!
-
--- Set highlight on search
-vim.o.hlsearch = false
--- Make line numbers default
-vim.wo.number = true
--- Enable mouse mode
-vim.o.mouse = 'a'
--- Sync clipboard between OS and Neovim.
---  Remove this option if you want your OS clipboard to remain independent.
---  See `:help 'clipboard'`
-vim.o.clipboard = 'unnamedplus'
--- Enable break indent
-vim.o.breakindent = true
--- Save undo history
-vim.o.undofile = true
--- Case insensitive searching UNLESS /C or capital in search
-vim.o.ignorecase = true
-vim.o.smartcase = true
--- Keep signcolumn on by default
-vim.wo.signcolumn = 'yes'
--- Decrease update time
-vim.o.updatetime = 250
-vim.o.timeout = true
-vim.o.timeoutlen = 300
-
--- Set completeopt to have a better completion experience
-vim.o.completeopt = 'menuone,noselect'
-
--- NOTE: You should make sure your terminal supports this
-vim.o.termguicolors = true
-vim.opt.relativenumber = true
-vim.opt.tabstop = 4
-vim.opt.softtabstop = 4
-vim.opt.shiftwidth = 4
-vim.opt.expandtab = true
-vim.opt.smartindent = true
-vim.opt.wrap = false
-
--- [[ Basic Keymaps ]]
-
--- Keymaps for better default experience
--- See `:help vim.keymap.set()`
-vim.keymap.set({ 'n', 'v' }, '<Space>', '<Nop>', { silent = true })
-
--- Remap for dealing with word wrap
-vim.keymap.set('n', 'k', "v:count == 0 ? 'gk' : 'k'", { expr = true, silent = true })
-vim.keymap.set('n', 'j', "v:count == 0 ? 'gj' : 'j'", { expr = true, silent = true })
-
--- move between splits
-vim.keymap.set('n', '<C-k>', ':wincmd k<CR>', { silent = true })
-vim.keymap.set('n', '<C-j>', ':wincmd j<CR>', { silent = true })
-vim.keymap.set('n', '<C-h>', ':wincmd h<CR>', { silent = true })
-vim.keymap.set('n', '<C-l>', ':wincmd l<CR>', { silent = true })
-
--- open diagnostics in qf list
-vim.keymap.set('n', '<leader>qf', ':lua vim.diagnostic.setqflist()<CR>', { silent = true })
-
--- [[ Highlight on yank ]]
--- See `:help vim.highlight.on_yank()`
-local highlight_group = vim.api.nvim_create_augroup('YankHighlight', { clear = true })
-vim.api.nvim_create_autocmd('TextYankPost', {
-  callback = function()
-    vim.highlight.on_yank()
-  end,
-  group = highlight_group,
-  pattern = '*',
-})
 
 -- [[ setup indent_blankline ]]
 require('ibl').setup()
@@ -404,11 +337,6 @@ require('nvim-treesitter.configs').setup {
 }
 vim.treesitter.language.register('templ', 'templ')
 
--- Diagnostic keymaps
-vim.keymap.set('n', '[d', vim.diagnostic.goto_prev, { desc = 'Go to previous diagnostic message' })
-vim.keymap.set('n', ']d', vim.diagnostic.goto_next, { desc = 'Go to next diagnostic message' })
-vim.keymap.set('n', '<leader>e', vim.diagnostic.open_float, { desc = 'Open floating diagnostic message' })
-vim.keymap.set('n', '<leader>q', vim.diagnostic.setloclist, { desc = 'Open diagnostics list' })
 
 -- [[ Configure LSP ]]
 --  This function gets run when an LSP connects to a particular buffer.
@@ -461,7 +389,7 @@ end
 --  Add any additional override configuration in the following tables. They will be passed to
 --  the `settings` field of the server config. You must look up that documentation yourself.
 local servers = {
-  bashls = {},
+  -- bashls = {},
   -- clangd = {},
   gopls = {},
   pyright = {},
@@ -474,8 +402,8 @@ local servers = {
     },
   },
   -- ruby_ls = {},
-  templ = {},
-  ocamllsp = {}
+  -- templ = {},
+  -- ocamllsp = {}
 }
 
 -- Setup neovim lua configuration
@@ -558,31 +486,3 @@ cmp.setup {
     { name = 'luasnip' },
   },
 }
-
--- dynamically set tabspaces
-local cmd = vim.cmd
-cmd [[
-  autocmd BufEnter *.py :setlocal tabstop=4 shiftwidth=4 expandtab
-  autocmd BufEnter *.js :setlocal tabstop=2 shiftwidth=2 expandtab
-  autocmd BufEnter *.ts :setlocal tabstop=2 shiftwidth=2 expandtab
-  autocmd BufEnter *.lua :setlocal tabstop=2 shiftwidth=2 expandtab
-  autocmd BufEnter *.tsx :setlocal tabstop=2 shiftwidth=2 expandtab
-]]
-
--- The line beneath this is called `modeline`. See `:help modeline`
--- vim: ts=2 sts=2 sw=2 et
---
--- controls for moving in and out of a terminal
-function _G.set_terminal_keymaps()
-  local opts = { buffer = 0 }
-  vim.keymap.set('t', '<esc>', [[<C-\><C-n>]], opts)
-  vim.keymap.set('t', 'jk', [[<C-\><C-n>]], opts)
-  vim.keymap.set('t', '<C-h>', [[<Cmd>wincmd h<CR>]], opts)
-  vim.keymap.set('t', '<C-j>', [[<Cmd>wincmd j<CR>]], opts)
-  vim.keymap.set('t', '<C-k>', [[<Cmd>wincmd k<CR>]], opts)
-  vim.keymap.set('t', '<C-l>', [[<Cmd>wincmd l<CR>]], opts)
-  vim.keymap.set('t', '<C-w>', [[<C-\><C-n><C-w>]], opts)
-end
-
--- if you only want these mappings for toggle term use term://*toggleterm#* instead
-vim.cmd('autocmd! TermOpen term://* lua set_terminal_keymaps()')
